fix followButton
